requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.8.1
    # via django
django==5.1.1
    # via
    #   -r .\requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r .\requirements.in
psycopg2-binary==2.9.9
    # via -r .\requirements.in
sqlparse==0.5.1
    # via django
tzdata==2024.1
    # via django
